🚚move: NavBar 파일, Toast 파일 Layout 폴더로 이동
