+++ grass-diary/src/components/Layout/Layout.tsx
@@ -1,7 +1,7 @@
-import Footer from '@components/NavBar/Footer';
-import Header from '@components/NavBar/Header';
-import Toast from '@components/Toast/Toast';
 import { Outlet } from 'react-router-dom';
+import Header from './Header';
+import Footer from './Footer';
+import Toast from './Toast';
 
 const Layout = () => {
   return (
