+++ grass-diary/src/components/index.ts
@@ -1,4 +1,3 @@
-export { default as Header } from './NavBar/Header';
 export { default as BackButton } from './Button/BackButton';
 export { default as Button } from './Button/Button';
 export { default as Like } from './Like/Like';
@@ -10,5 +9,5 @@ export { default as MoodProfile } from './Profile/MoodProfile';
 export { default as ProtectedRoute } from './ProtectedRoute';
 export { default as Container } from './Container';
 export { EllipsisIcon, EllipsisBox } from './Ellipsis';
-export { default as Footer } from './NavBar/Footer';
 export { default as Comments } from './Comment/Comments';
+export { default as Layout } from './Layout/Layout';
